Moved LayerManager component below TheMap component.
