+++ frontend/src/Components/MapContextForm/MapContext.tsx
@@ -13,8 +13,8 @@ import MapContextLayerRepo from '../../Repos/MapContextLayerRepo';
 import MapContextRepo from '../../Repos/MapContextRepo';
 import { LayerUtils } from '../../Utils/LayerUtils';
 import { TreeUtils } from '../../Utils/TreeUtils';
-import { CreateLayerOpts } from '../LayerManager/LayerManagerTypes';
 import { TreeFormFieldDropNodeEventType, TreeNodeType } from '../Shared/FormFields/TreeFormField/TreeFormFieldTypes';
+import { CreateLayerOpts } from '../TheMap/LayerManager/LayerManagerTypes';
 import { olMap, TheMap } from '../TheMap/TheMap';
 import './MapContext.css';
 import { MapContextForm } from './MapContextForm';
