+++ frontend/src/Components/TheMap/LayerManager/LayerManager.tsx
@@ -13,13 +13,13 @@ import { transformExtent } from 'ol/proj';
 import ImageWMS from 'ol/source/ImageWMS';
 import { getUid } from 'ol/util';
 import React, { useEffect, useState } from 'react';
-import { JsonApiResponse } from '../../Repos/JsonApiRepo';
-import LayerRepo from '../../Repos/LayerRepo';
-import { LayerManagerUtils } from '../../Utils/LayerManagerUtils';
-import { LayerUtils } from '../../Utils/LayerUtils';
-import { TreeUtils } from '../../Utils/TreeUtils';
-import { TreeFormField } from '../Shared/FormFields/TreeFormField/TreeFormField';
-import { TreeFormFieldDropNodeEventType, TreeNodeType } from '../Shared/FormFields/TreeFormField/TreeFormFieldTypes';
+import { JsonApiResponse } from '../../../Repos/JsonApiRepo';
+import LayerRepo from '../../../Repos/LayerRepo';
+import { LayerManagerUtils } from '../../../Utils/LayerManagerUtils';
+import { LayerUtils } from '../../../Utils/LayerUtils';
+import { TreeUtils } from '../../../Utils/TreeUtils';
+import { TreeFormField } from '../../Shared/FormFields/TreeFormField/TreeFormField';
+import { TreeFormFieldDropNodeEventType, TreeNodeType } from '../../Shared/FormFields/TreeFormField/TreeFormFieldTypes';
 import './LayerManager.css';
 import { CreateLayerOpts, LayerManagerProps } from './LayerManagerTypes';
