+++ frontend/src/Utils/TreeUtils.ts
@@ -3,8 +3,8 @@ import BaseLayer from 'ol/layer/Base';
 import LayerGroup from 'ol/layer/Group';
 import ImageLayer from 'ol/layer/Image';
 import ImageWMS from 'ol/source/ImageWMS';
-import { CreateLayerOpts } from '../Components/LayerManager/LayerManagerTypes';
 import { MPTTJsonApiTreeNodeType, TreeNodeType } from '../Components/Shared/FormFields/TreeFormField/TreeFormFieldTypes';
+import { CreateLayerOpts } from '../Components/TheMap/LayerManager/LayerManagerTypes';
 import { LayerUtils } from './LayerUtils';
 
 const layerUtils = new LayerUtils();
